requirements.txt
# Core dependencies
torch>=2.0.0
torchvision
diffusers>=0.35.0
transformers>=4.51.3
accelerate>=0.21.0
peft>=0.7.0

# RunPod SDK
runpod

# Image handling
pillow
safetensors

# Optional: For downloading models at runtime
huggingface_hub

# Cloud storage (R2/S3)
boto3>=1.26.0
python-dotenv>=0.19.0
